Move exec out of terminalTab and into terminalActions to resolve code layering issue
